+++ R/verb-head.R
@@ -64,5 +64,5 @@ add_head <- function(x, n) {
 
 #' @export
 tail.tbl_lazy <- function(x, n = 6L, ...) {
-  stop("tail() is not supported by sql sources", call. = FALSE)
+  abort("tail() is not supported by sql sources")
 }