+++ R/verb-pivot-wider.R
@@ -76,10 +76,10 @@ pivot_wider.tbl_lazy <- function(data,
                                  names_repair = "check_unique",
                                  values_from = value,
                                  values_fill = NULL,
-                                 values_fn = max,
+                                 values_fn = ~ max(.x, na.rm = TRUE),
                                  ...
                                  ) {
-  ellipsis::check_dots_empty()
+  rlang::check_dots_empty()
   names_from <- enquo(names_from)
   values_from <- enquo(values_from)
 
@@ -146,19 +146,8 @@ dbplyr_pivot_wider_spec <- function(data,
                                     names_repair = "check_unique",
                                     id_cols = NULL,
                                     values_fill = NULL,
-                                    values_fn = max) {
+                                    values_fn = ~ max(.x, na.rm = TRUE)) {
   spec <- check_spec(spec)
-
-  if (is.null(values_fn)) {
-    abort(c(
-      "`values_fn` must not be NULL",
-      i = "`values_fn` must be a function or a named list of functions"
-    ))
-  }
-  if (is.function(values_fn)) {
-    values_fn <- rep_named(unique(spec$.value), list(values_fn))
-    values_fn <- purrr::map_chr(values_fn, find_fun)
-  }
 
   if (is_scalar(values_fill)) {
     values_fill <- rep_named(unique(spec$.value), list(values_fill))
@@ -170,6 +159,23 @@ dbplyr_pivot_wider_spec <- function(data,
   values <- vctrs::vec_unique(spec$.value)
   spec_cols <- c(names(spec)[-(1:2)], values)
 
+  if (is.null(values_fn)) {
+    abort(c(
+      "`values_fn` must not be NULL",
+      i = "`values_fn` must be a function or a named list of functions"
+    ))
+  }
+
+  if (!vctrs::vec_is_list(values_fn)) {
+    values_fn <- rep_named(values, list(values_fn))
+  }
+
+  values_fn <- purrr::compact(values_fn)
+  missing_values <- setdiff(values, names(values_fn))
+  if (!is_empty(missing_values)) {
+    abort("`values_fn` must specify a function for each col in `values_from`")
+  }
+
   id_cols <- enquo(id_cols)
   if (!quo_is_null(id_cols)) {
     cn <- set_names(colnames(data))
@@ -179,18 +185,26 @@ dbplyr_pivot_wider_spec <- function(data,
   }
   key_vars <- setdiff(key_vars, spec_cols)
 
-  key_col <- sym(names(spec)[3])
+  key_cols <- syms(names(spec)[-(1:2)])
   pivot_exprs <- purrr::map(
     vctrs::vec_seq_along(spec),
     function(row) {
-      key <- spec[[3]][row]
       values_col <- spec[[".value"]][row]
-
       fill_value <- values_fill[[values_col]]
-      case_expr <- expr(ifelse(!!key_col == !!key, !!sym(values_col), !!fill_value))
+
+      keys <- vctrs::vec_slice(spec[, -(1:2)], row)
+      keys_cond <- purrr::imap(
+        keys,
+        function(value, name) {
+          expr(!!sym(name) == !!value)
+        }
+      ) %>%
+        purrr::reduce(~ expr(!!.x & !!.y))
+
+      case_expr <- expr(ifelse(!!keys_cond, !!sym(values_col), !!fill_value))
 
       agg_fn <- values_fn[[values_col]]
-      expr((!!agg_fn)(!!case_expr, na.rm = TRUE))
+      resolve_fun(agg_fn, case_expr)
     }
   ) %>%
     set_names(spec$.name)
@@ -222,5 +236,16 @@ is_scalar <- function(x) {
     (length(x) == 1) && !have_name(x)
   } else {
     length(x) == 1
+  }
+}
+
+
+resolve_fun <- function(x, var) {
+  if (is_formula(x)) {
+    .fn_expr <- across_fun(x)
+    exec(.fn_expr, var)
+  } else {
+    fn_name <- find_fun(x)
+    call2(fn_name, var)
   }
 }